requirements.txt
flask
flask-session
bcrypt
requests
python-dotenv
uuid
flask-wtf
wtforms